refactor(falabella): merge CMR credit card scraper into Banco Falabella
CMR Falabella is Banco Falabella's own credit card — same portal, same
credentials. Merging them into one scraper matches the pattern of BCI,
Itaú and bchile which all return account + TC movements together.

- falabella scraper now returns:
  - Account (cuenta corriente) movements
  - [TC Por Facturar] movements (from CMR Shadow DOM)
  - [TC Facturados] movements (from CMR Shadow DOM)
  - creditCards[] with CMR cupos (national: total/used/available)
- --owner T|A|B flag now works for Falabella TC filtering
- Remove src/banks/cmr.ts (logic merged into falabella.ts)
- Remove cmr from index.ts bank registry
- Update .env.example: remove CMR_RUT/CMR_PASS (use FALABELLA_* instead)
- Update README: one row for Banco Falabella with "(cuenta + CMR TC)"
