+++ .env.example
@@ -1,6 +1,5 @@
 # Tu RUT sin puntos ni guión (ej: 123456789)
 FALABELLA_RUT=123456789
-CMR_RUT=123456789
 BICE_RUT=123456789
 SANTANDER_RUT=123456789
 EDWARDS_RUT=123456789
@@ -12,7 +11,6 @@ SCOTIABANK_RUT=123456789
 
 # Tu clave de internet del banco
 FALABELLA_PASS=tu_clave_aqui
-CMR_PASS=tu_clave_aqui
 BICE_PASS=tu_clave_aqui
 SANTANDER_PASS=tu_clave_aqui
 EDWARDS_PASS=tu_clave_aqui
@@ -21,6 +19,9 @@ ITAU_PASS=tu_clave_aqui
 BANCOCHILE_PASS=tu_clave_aqui
 BCHILE_PASS=tu_clave_aqui
 SCOTIABANK_PASS=tu_clave_aqui
+
+# Opcional [FALABELLA]: filtro TC por titular/adicional (T=titular, A=adicional, B=todos, default: B)
+# FALABELLA_OWNER=B
 # Opcional [SCOTIABANK]:
 # SCOTIABANK_MONTHS=3
 # Opcional [BCHILE]: segundos máximos para esperar aprobación manual de 2FA (default: 180)
